+++ lizard_fancylayers/tests/test_layers.py
@@ -38,25 +38,6 @@ class TestAdapter(TestCase):
 
         self.assertTrue(html)
 
-    def test_returns_legend(self, patched_datasource):
-        workspace_item = mock.MagicMock()
-
-        adapter = layers.FancyLayersAdapter(
-            workspace_item,
-            layer_arguments={'choices_made': "{}"})
-
-        legend = adapter.legend()
-
-        # Must be an iterable
-        l = list(legend)
-
-        # Must be something in it
-        self.assertTrue(len(l) > 0)
-
-        # Must have right keys
-        self.assertTrue('img_url' in l[0])
-        self.assertTrue('description' in l[0])
-
     def test_empty_legend(self, patched_datasource):
         # Don't return a legend if there are no annotations to grab colors
         # from.